requirements.txt
reflex==0.6.5
python-decouple
openai
